refactor : 도메인 변경 적용으로 인한 import 최적화
